Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,14 +1,11 @@
 #include <stdio.h>
 
 
-
-
-
 int main() {
     int a, b, c;
     scanf("%i", a);
     scanf("%i", b);
     scanf("%i", c);
-    printf("Average: %i", (a+b+c)/3);
+    printf("Average: %.2f", (float) (a+b+c)/3);
     return 0;
 }
